% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_multiple_tables.R
\name{make_multiple_tables}
\alias{make_multiple_tables}
\title{Make multiple short tables out of one long table}
\usage{
make_multiple_tables(df, row_length = 30)
}
\arguments{
\item{df}{a very long table}

\item{row_length}{the amount of rows per each table}
}
\value{

}
\description{
Make multiple short tables out of one long table
}
\examples{

}
