% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_vector_to_SQL_vector.R
\name{R_vector_to_SQL_vector}
\alias{R_vector_to_SQL_vector}
\title{A function used for passing a vector in R
to become a string in SQL.}
\usage{
R_vector_to_SQL_vector(v)
}
\arguments{
\item{v}{a vector to be converted}
}
\value{

}
\description{
A function used for passing a vector in R
to become a string in SQL.
}
